// src/generated/models/PowerPlatformforAdminsV2Model.ts
export interface AdvisorRecommendationResource {
  // The resource display name
  resourceName?: string;
  // The resource unique ID
  resourceId?: string;
  // The resource owner object ID
  resourceOwnerId?: string;
  // The resource owner display name
  resourceOwner?: string;
  // The type of resource
  resourceType?: string;
  // The sub type of the resource
  resourceSubType?: string;
  // The resource description
  resourceDescription?: string;
  // Number of unique users who used the resource in the last thirty (30) days
  resourceUsage?: number;
  // The environment display name
  environmentName?: string;
  // The environment unique ID
  environmentId?: string;
  // Time when the resource was last modified
  lastModifiedDate?: string;
  // Time when the resource was last used
  lastAccessedDate?: string;
  // Current status of any action taken since the last refresh time
  resourceActionStatus?: string;
}